fix: resolve CI failures — add missing mock and extract MediaLibraryPanel
- Add `toDirectSignedUrl` to video-worker test mock (fixes unit test)
- Extract MediaLibraryPanel from VideoEditorStage (521→343 lines, fixes guard)
